refactor structure to enable dali, tar and jpegtran cpabilities ... unfinished
